# perf: remove std::optional creation for host parsing