--- a/include/ada/implementation.h
+++ b/include/ada/implementation.h
@@ -33,13 +33,13 @@ namespace ada {
                                  std::optional<ada::state> state = std::nullopt) noexcept;
 
   void set_scheme(ada::url &base, std::string input, ada::encoding_type encoding = ada::encoding_type::UTF8) noexcept;
-  void set_username(ada::url &base, std::string input) noexcept;
-  void set_password(ada::url &base, std::string input) noexcept;
+  void set_username(ada::url &base, std::string_view input) noexcept;
+  void set_password(ada::url &base, std::string_view input) noexcept;
   void set_host(ada::url &base, std::string_view input, ada::encoding_type encoding = ada::encoding_type::UTF8) noexcept;
   void set_port(ada::url &base, std::string_view input) noexcept;
-  void set_pathname(ada::url &base, std::string input, ada::encoding_type encoding = ada::encoding_type::UTF8) noexcept;
-  void set_search(ada::url &base, std::string input) noexcept;
-  void set_hash(ada::url &base, std::string input) noexcept;
+  void set_pathname(ada::url &base, std::string_view input, ada::encoding_type encoding = ada::encoding_type::UTF8) noexcept;
+  void set_search(ada::url &base, std::string_view input) noexcept;
+  void set_hash(ada::url &base, std::string_view input) noexcept;
 }
 
 #endif // ADA_IMPLEMENTATION_H