--- a/include/ada/parser.h
+++ b/include/ada/parser.h
@@ -12,9 +12,10 @@ namespace ada::parser {
   // first_percent should be  = plain.find('%')
   std::optional<std::string> to_ascii(std::string_view plain, bool be_strict, size_t first_percent);
 
-  std::optional<ada::url_host> parse_opaque_host(std::string_view input);
-  std::optional<ada::url_host> parse_ipv6(std::string_view input);
-  std::optional<ada::url_host> parse_host(std::string_view input, bool is_not_special, bool input_is_ascii);
+  bool parse_opaque_host(std::optional<ada::url_host>& out, std::string_view input);
+  bool parse_ipv6(std::optional<ada::url_host>& out, std::string_view input);
+  bool parse_ipv4(std::optional<ada::url_host>& out, std::string_view input);
+  bool parse_host(std::optional<ada::url_host>& out, std::string_view input, bool is_not_special, bool input_is_ascii);
 
   url parse_url(std::string_view user_input,
                 std::optional<ada::url> base_url = std::nullopt,