improves parsing by around ~0.32ns per byte.

before

```
BasicBench_AdaURL       5097 ns         5093 ns       104205 time/byte=6.20336ns time/url=463.01ns url/s=2.15978M/s
```

after

```
BasicBench_AdaURL       4775 ns         4775 ns       107061 time/byte=5.90992ns time/url=434.11ns url/s=2.30356M/s
```