@lemire 2023-01-16
I am preparing a PR on top of this PR. I share some of @addaleax 's concerns.